remove block delta & rename state to catchup
