+++ sequencer/src/api/data_source.rs
@@ -3,7 +3,12 @@ use super::{
     options::{Options, Query},
     sql,
 };
-use crate::{network, persistence, state::ValidatedState, Node, SeqTypes};
+use crate::{
+    network, persistence,
+    state::{BlockMerkleTree, Delta, FeeAccount, FeeMerkleTree, ValidatedState},
+    Node, SeqTypes,
+};
+use anyhow::Context;
 use async_std::sync::Arc;
 use async_trait::async_trait;
 use hotshot::types::SystemContextHandle;
@@ -11,12 +16,15 @@ use hotshot_query_service::{
     availability::AvailabilityDataSource,
     data_source::{UpdateDataSource, VersionedDataSource},
     fetching::provider::{AnyProvider, QueryServiceProvider},
-    merklized_state::MerklizedState,
+    merklized_state::{MerklizedState, UpdateStateStorage},
     node::NodeDataSource,
     status::StatusDataSource,
+    Leaf,
 };
 use hotshot_types::{data::ViewNumber, light_client::StateSignatureRequestBody};
-use jf_primitives::merkle_tree::prelude::MerklePath;
+use jf_primitives::merkle_tree::{
+    prelude::MerklePath, MerkleTreeScheme, ToTraversalPath, UniversalMerkleTreeScheme,
+};
 use tide_disco::Url;
 use versioned_binary_serialization::version::StaticVersionType;
 
@@ -97,6 +105,59 @@ pub(crate) trait StateSignatureDataSource<N: network::Type> {
 pub(crate) trait LocalStateDataSource {
     async fn get_decided_state(&self) -> Arc<ValidatedState>;
     async fn get_undecided_state(&self, view: ViewNumber) -> Option<Arc<ValidatedState>>;
+}
+
+#[async_trait]
+impl<D: SequencerDataSource + Send + Sync> UpdateStateStorage<SeqTypes, D> for ValidatedState {
+    async fn update_storage(
+        &self,
+        storage: &mut D,
+        leaf: &Leaf<SeqTypes>,
+        delta: Arc<Delta>,
+    ) -> anyhow::Result<()> {
+        let block_number = leaf.get_height();
+        let ValidatedState {
+            fee_merkle_tree,
+            block_merkle_tree,
+        } = self;
+
+        let Delta { fees_delta } = delta.as_ref();
+
+        // Insert block merkle tree nodes
+        let (_, proof) = block_merkle_tree
+            .lookup(block_number - 1)
+            .expect_ok()
+            .context("Index not found in block merkle tree")?;
+        let path = <u64 as ToTraversalPath<typenum::U3>>::to_traversal_path(
+            &(block_number - 1),
+            block_merkle_tree.height(),
+        );
+
+        storage
+            .store_state::<BlockMerkleTree>(proof.proof, path, block_number)
+            .await
+            .context("failed to insert merkle nodes for block merkle tree")?;
+
+        // Insert fee merkle tree nodes
+        for delta in fees_delta {
+            let (_, proof) = fee_merkle_tree
+                .universal_lookup(delta)
+                .expect_ok()
+                .context("Index not found in fee merkle tree")?;
+            let path: Vec<usize> =
+                <FeeAccount as ToTraversalPath<typenum::U256>>::to_traversal_path(
+                    delta,
+                    fee_merkle_tree.height(),
+                );
+
+            storage
+                .store_state::<FeeMerkleTree>(proof.proof, path, block_number)
+                .await
+                .context("failed to insert merkle nodes for block merkle tree")?;
+        }
+
+        Ok(())
+    }
 }
 
 #[cfg(test)]
